% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TH_field_code.R
\name{download}
\alias{download}
\title{Download from SD to Laptop}
\usage{
download(input = input.drive, output = output.folder)
}
\description{
Downloads TreeHugger .CSV from chosen drive to a desktop folder. Relies on prompted user input to determine appropriate download and upload paths. Warns user if download or upload conditions are not as expected. Outputs the copied file name to the console.
}
