% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TH_field_code.R
\name{set.folders}
\alias{set.folders}
\title{Set input/output folder for TH download/upload}
\usage{
set.folders()
}
\description{
This function sets paths for the set() and download() functions and should be used at the beginning of a session. The assigned input drive and output folder are printed in the R console. Please create an download output folder specific to the download day and site. Note: No parameters are necessary.
}
