% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TH_field_code.R
\name{set}
\alias{set}
\title{Logger settings}
\usage{
set(name = "N", int = 30, logger = "new")
}
\arguments{
\item{name}{The logger ID. A five digit number. For "old" loggers (purple board), the convention is 14XXX. For "new" loggers (green board), the convention is 16XXX.}

\item{int}{The logging interval; a number in minutes. Typically, do not change this from 30.}

\item{logger}{'old' vs 'new' Old= a purple board, New = a green board. Differentiation is necessary because the logger versions require different time file names and time formats.}
}
\value{
Up to three different setting files to the selected removeable drive depending on user inputs.
}
\description{
Sets key logger files: logging interval, time, logger ID. Setting the logger time is automatic (and will be to non-daylight savings time). Setting the logging interval and changing the ID will require USER responses. Note: In the current version, you can set parameter values for each of the logger setting files; alternatively, you can set these values with the prompts that result from running the function w/o parameters. Following the prompts will override any parameters set initially.
}
\examples{
set("16999", 30, "new")

}
