% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TH_field_code.R
\name{set.folders}
\alias{set.folders}
\title{Set input/output folder for TH download/upload}
\usage{
set.folders()
}
\description{
This function sets paths for the set() and download() functions and should be used at the beginning of a session
}
